+++ plugins/data-management/src/AddTrackWidget/components/TrackSourceSelect.tsx
@@ -1,7 +1,6 @@
 import { FileSelector } from '@jbrowse/core/ui'
 import Paper from '@material-ui/core/Paper'
 import { makeStyles } from '@material-ui/core/styles'
-import PropTypes from 'prop-types'
 import React from 'react'
 import { AddTrackModel } from '../model'
 
@@ -35,13 +34,6 @@ function TrackSourceSelect({ model }: { model: AddTrackModel }) {
       </Paper>
     </div>
   )
-}
-
-TrackSourceSelect.propTypes = {
-  trackData: PropTypes.objectOf(PropTypes.any).isRequired,
-  indexTrackData: PropTypes.objectOf(PropTypes.any).isRequired,
-  setTrackData: PropTypes.func.isRequired,
-  setIndexTrackData: PropTypes.func.isRequired,
 }
 
 export default TrackSourceSelect
